+++ kurdikova.elizaveta/I2/ShiftArrayElements.h
@@ -1,6 +1,6 @@
 #ifndef SHIFTARRAYELEMENTS_H
 #define SHIFTARRAYELEMENTS_H
 
-bool shiftArrayElements(int* arr, size_t size, int shift);
+bool shiftArrayElements(int* arr, size_t size, size_t shift);
 
 #endif
